+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Braintree Android SDK Release Notes
 
+## unreleased
+* Add `isExplicitCancellation` parameter to `UserCanceledException`
+
 ## 4.10.1
 
 * DataCollector