+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Braintree Android SDK Release Notes
 
 ## unreleased
-* Add `isExplicitCancellation` parameter to `UserCanceledException`
+* Add `isExplicitCancelation` parameter to `UserCanceledException`
 
 ## 4.10.1